Rename: CSS Styl5t Ed3r code for WSUWP Spine theme
Rename CSS Stylesheet Editor code file «CSS/lsamp-custom.prev.css» to «CSS/wsuwp-spine_css-styl5t-ed3r.prod5n.css» in anticipation of upgrading to the WDS theme for WordPress on the LSAMP website. Use this new file naming pattern to include the name of the WordPress theme for which the production code build was designed.